Modulize agg_index_reader.
