+++ src/query/storages/fuse/src/io/read/agg_index/mod.rs
@@ -13,5 +13,7 @@
 // limitations under the License.
 
 mod agg_index_reader;
+mod agg_index_reader_native;
+mod agg_index_reader_parquet;
 
 pub use agg_index_reader::AggIndexReader;
