+++ src/query/storages/fuse/src/io/read/block/mod.rs
@@ -25,4 +25,5 @@ mod decompressor;
 pub use block_reader::BlockReader;
 pub use block_reader_merge_io::MergeIOReadResult;
 pub use block_reader_native::NativeReaderExt;
+pub use block_reader_native::NativeSourceData;
 pub use decompressor::UncompressedBuffer;
